Add gemfile and install rubocop, add code for filter method in my_enumerable, and corrects linting errors
