+++ Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
+
+ruby '3.1.2'
+
+gem 'rubocop', '>= 1.0', '< 2.0'
